Added carbon filter and apply button
