+++ src/components/mapsComponents/SearchBar.css
@@ -1,8 +1,3 @@
-#search-bar {
-  display: flex;
-  flex-direction: column;
-  width: 20rem;
-}
 #main-search {
   color: #333333;
   height: 2.5rem;
@@ -28,93 +23,4 @@
 }
 #main-search img {
   cursor: pointer;
-}
-#suggestions {
-  max-height: 240px;
-  min-height: 39px;
-  width: 100%;
-  color: #333333;
-  background-color: #fefefe;
-  box-shadow: 0 0 5px 0 #999999;
-  padding: 5px 0;
-  border-radius: 4px;
-  font-weight: 300;
-  overflow-y: auto;
-  margin-top: 5px;
-  display: none;
-}
-#suggestions p {
-  padding: 5px 20px;
-  cursor: pointer;
-  font-size: 0.9rem;
-}
-#suggestions p:hover {
-  background-color: #eeeeee;
-}
-.place-name-dropdown {
-  font-weight: 400;
-  font-size: 0.9rem;
-}
-.place-name-dropdown.highlight {
-  background-color: #eeeeee;
-}
-.state-nation-name-dropdown {
-  font-size: 0.7rem;
-  color: #bdbdbd;
-  display: block;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-#confirmation {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background-color: rgba(10, 10, 10, 0.2);
-  z-index: 1;
-  justify-content: center;
-  align-items: center;
-  display: none;
-}
-#confirmation-panel {
-  height: 400px;
-  width: 300px;
-  color: #333333;
-  background-color: #fefefe;
-  box-shadow: 0 0 5px 0 #999999;
-  border-radius: 4px;
-  padding: 20px;
-  overflow-y: auto;
-}
-#confirmation-panel .close-btn {
-  float: right;
-  cursor: pointer;
-}
-#confirmation-panel .confirmation-title {
-  font-size: 1.4rem;
-  font-weight: 400;
-  margin-bottom: 20px;
-}
-#confirmation-panel .confirmation-item {
-  margin-bottom: 1rem;
-  cursor: pointer;
-}
-#confirmation-panel .confirmation-item:hover {
-  color: #000000;
-  font-weight: 400;
-}
-
-@media (max-width: 959px) and (min-width: 768px) {
-  #search-bar {
-    width: 18rem;
-  }
-}
-
-@media (max-width: 767px) {
-  #search-bar {
-    width: calc(100vw - 140px);
-    max-width: 20rem;
-  }
 }
