+++ src/components/mapsComponents/SearchBar.js
@@ -1,172 +1,40 @@
 import React, { useState, useEffect } from "react";
-import useDebounce from "./helpers/useDebounce";
 import searchIcon from "../../images/search.svg";
-import closeIcon from '../../images/x.svg';
 import "./SearchBar.css";
 
 export default function SearchBar({
-  getSimilarPlaces,
-  filters,
-  targetPlaceID,
-  setTargetPlace,
+  searchTerm,
+  setSearchTerm,
+  handlePlaceUpdates,
 }) {
-  const [searchTerm, setSearchTerm] = useState("");
-  const debouncedSearchTerm = useDebounce(searchTerm, 1000);
-  const [isSearching, setIsSearching] = useState(false);
-  const [placeSuggestions, setPlaceSuggestions] = useState([]);
-  const [selectedSuggestion, setSelectedSuggestion] = useState([]);
-  useEffect(() => {
-    if (debouncedSearchTerm && debouncedSearchTerm.length > 1) {
-      setIsSearching(true);
-      fetch(
-        `http://localhost:8080/api/places/${debouncedSearchTerm}`
-      )
-        .then(response => response.json())
-        .then(results => {
-          if (results.features) {
-            setPlaceSuggestions(results.features);
-            setSelectedSuggestion([
-              results.features[0].properties.place_id,
-              0,
-            ]);
-          } else {
-            setPlaceSuggestions([]);
-            setSelectedSuggestion([]);
-          }
-          setIsSearching(false);
-        });
-    } else {
-      setPlaceSuggestions([]);
-      setSelectedSuggestion([]);
-    }
-  }, [debouncedSearchTerm]);
-
-  const openConfirmationPanel = () => {
-    document.getElementById('suggestions').style.display = 'none';
-    document.getElementById('confirmation').style.display = 'flex';
-  }
-  const closeConfirmationPanel = () => {
-    document.getElementById('confirmation').style.display = 'none';
-  }
-  const handlePlaceUpdates = () => {
-    if (isSearching) {
-      return;
-    }
-    if (placeSuggestions.length > 1) {
-      openConfirmationPanel();
-    } else if (placeSuggestions.length === 1) {
-      setSearchTerm(placeSuggestions[0].properties.name);
-      setSelectedSuggestion([
-        placeSuggestions[0].properties.place_id,
-        0,
-      ]);
-      setTargetPlace(placeSuggestions[0]);
-      getSimilarPlaces(
-        placeSuggestions[0].properties.place_id,
-        filters
-      );
-      document.getElementById('suggestions').style.display = 'none';
-    }
-  }
   const handleKeyDown = (event) => {
     if (event.key === 'Enter') {
       handlePlaceUpdates();
     }
   };
-  const handleSuggestionClick = (placeID, name, index) => {
-    setSearchTerm(name);
-    if (placeID !== targetPlaceID) {
-      setSelectedSuggestion([placeID, index]);
-      setTargetPlace(placeSuggestions[index]);
-      getSimilarPlaces(placeID, filters);
-    }
-  };
-
   return (
-    <div id="search-bar">
-      <div id="main-search">
-        <input
-          type="text"
-          placeholder="Search by place name"
-          value={searchTerm}
-          onChange={event => setSearchTerm(event.target.value)}
-          onKeyDown={event => handleKeyDown(event)}
-          onFocus={() => {
-            document.querySelector("#suggestions").style.display = "block";
-          }}
-          onBlur={() => {
-            setTimeout(() => {
-              document.querySelector("#suggestions").style.display = "none";
-            }, 200);
-          }}
-        />
-        <img
-          src={searchIcon}
-          alt="search"
-          id="search"
-          onClick={() => handlePlaceUpdates()}
-        />
-      </div>
-      <div id="suggestions">
-        {isSearching ? (
-          <p>Searching...</p>
-        ) : placeSuggestions.length > 0 ? (
-          <>
-            {placeSuggestions.map(({ properties }, index) => {
-              const { place_id, name, state_name, nation_name } = properties;
-              return (
-                <p
-                  className={`place-name-dropdown${place_id === selectedSuggestion[0] ? ' highlight' : ''}`}
-                  key={place_id}
-                  onClick={() => {
-                    handleSuggestionClick(place_id, name, index);
-                  }}
-                >
-                  {name}
-                  <span className="state-nation-name-dropdown">
-                    {state_name} {state_name ? "," : ""} {nation_name}
-                  </span>
-                </p>
-              );
-            })}
-          </>
-        ) : debouncedSearchTerm.length === 1 ? (
-          <p>Please enter more than 1 letter</p>
-        ) : debouncedSearchTerm.length > 0 ? (
-          <p>No suggestion</p>
-        ) : (
-          <p>User Search History</p>
-        )}
-      </div>
-      <div id="confirmation">
-        <div id="confirmation-panel">
-          <img
-            className="close-btn"
-            src={closeIcon}
-            alt="close suggestion"
-            onClick={closeConfirmationPanel}
-          />
-          <p className="confirmation-title">Did you mean</p>
-          {placeSuggestions.map(({ properties }, index) => {
-            const { place_id, name, state_name, nation_name } = properties;
-            return (
-              <div
-                className="confirmation-item"
-                key={place_id}
-                onClick={() => {
-                  handleSuggestionClick(place_id, name, index);
-                  closeConfirmationPanel();
-                }}
-              >
-                <p>{name}</p>
-                <p className="state-nation-name-dropdown">
-                  {state_name} {state_name ? "," : ""} {nation_name}
-                </p>
-              </div>
-            );
-          })}
-        </div>
-      </div>
+    <div id="main-search">
+      <input
+        type="text"
+        placeholder="Search by place name"
+        value={searchTerm}
+        onChange={event => setSearchTerm(event.target.value)}
+        onKeyDown={event => handleKeyDown(event)}
+        onFocus={() => {
+          document.querySelector("#suggestions").style.display = "block";
+        }}
+        onBlur={() => {
+          setTimeout(() => {
+            document.querySelector("#suggestions").style.display = "none";
+          }, 200);
+        }}
+      />
+      <img
+        src={searchIcon}
+        alt="search"
+        id="search"
+        onClick={() => handlePlaceUpdates()}
+      />
     </div>
   );
-}
+};
