+++ src/components/mapsComponents/SuggestionDropdown.css
@@ -1,0 +1,37 @@
+#suggestions {
+  max-height: 240px;
+  min-height: 39px;
+  width: 100%;
+  color: #333333;
+  background-color: #fefefe;
+  box-shadow: 0 0 5px 0 #999999;
+  padding: 5px 0;
+  border-radius: 4px;
+  font-weight: 300;
+  overflow-y: auto;
+  margin-top: 5px;
+  display: none;
+}
+#suggestions p {
+  padding: 5px 20px;
+  cursor: pointer;
+  font-size: 0.9rem;
+}
+#suggestions p:hover {
+  background-color: #eeeeee;
+}
+.place-name-dropdown {
+  font-weight: 400;
+  font-size: 0.9rem;
+}
+.place-name-dropdown.highlight {
+  background-color: #eeeeee;
+}
+.state-nation-name-dropdown {
+  font-size: 0.7rem;
+  color: #bdbdbd;
+  display: block;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
